requirements.txt
# ===== Core Bot Framework =====
python-telegram-bot[webhooks,job-queue]

# ===== Environment & Configuration =====
python-dotenv

# ===== Networking =====
httpx[http2]

# ===== Database Connector =====
SQLAlchemy>=2.0
alembic==1.13.1
aiomysql
# Required by aiomysql for modern MySQL authentication
cryptography
alembic

# ===== Date & Time =====
jdatetime

# ===== Utilities =====
qrcode
Pillow
aiofiles

pytz==2024.1
PyMySQL